requirements.txt
# === Agent + LLM Tools ===
aisuite==0.1.11
anthropic
docstring-parser
markdown
mistralai
openai~=2.14.0
qrcode
tavily-python>=0.7.12
textstat
vertexai
grok

# === Web Framework + API ===
fastapi
pydantic
pydantic[email]
python-dotenv~=1.2.1
python-multipart
requests~=2.32.5
sqlalchemy
uvicorn


# === Notebook Experience ===
ipywidgets
jupyter_server
nbclassic
notebook

# === Data Analysis / Display (Optional Enhancements) ===
duckdb
matplotlib
pandas
seaborn
tabulate
tinydb

# === Machine Learning / NLP (Optional Enhancements) ===
jinja2
psycopg2-binary
scikit-learn
Wikipedia
groq~=1.0.0
protobuf~=5.29.5
google-auth-oauthlib~=1.2.3
google-api-python-client~=2.187.0
utils~=1.0.2
langchain-openai~=1.1.6
langchain-core~=1.2.4
pillow~=12.0.0